chore(package.json): update dependencies and scripts
This commit updates the dependencies and scripts in the package.json
file. The node engine version is updated to 18.0, and the dependencies
are updated to the latest versions. The cdktf version is updated to
0.16.x, and the provider-google version is updated to 6.x. The main and
types fields are updated to point to the correct file paths. The
test-environment.ts and setup-jest-cdktf.js files are moved to the src
directory.