feat(command_options.ts): add options for logging bucket name, analytics, and retention time
This commit adds three new options to the command line interface:
log-bucket, analytics, and retention. The log-bucket option allows the
user to specify the name of the logging bucket. The analytics option is
a boolean that determines whether analytics will be on or off for the
logging bucket. The retention option allows the user to specify the
retention time of logs in the bucket.