feat(license): add BSD 2-Clause License
This commit adds the BSD 2-Clause License to the project. The license
allows for the redistribution and use of the software in source and
binary forms, with or without modification, provided that the conditions
outlined in the license are met. The license also includes a disclaimer
of liability for any damages caused by the use of the software.