feat(.kodiak.toml): add Kodiak configuration file
This commit adds a new file named `.kodiak.toml` which contains
configuration settings for Kodiak. Kodiak is a GitHub bot that automates
code reviews and merges. The configuration file specifies that PRs with
the label "dependencies" and "automerge" will be automatically merged if
they are opened by the "dependabot" user and are "minor" or "patch"
version upgrades. The `ignored_usernames` setting is also added to the
`update` section to allow dependabot to update and close stale
dependency upgrades.