feat(dependabot.yml): add Dependabot configuration file
This commit adds a new file, `.github/dependabot.yml`, which contains
the configuration for Dependabot. Dependabot is a tool that automates
the process of keeping dependencies up to date. The configuration
specifies that Dependabot should check for updates to npm packages and
GitHub Actions every week on Saturday at 11:00 AM in the America/Chicago
timezone. The updates will be applied to the `develop` branch and will
create a maximum of 99 open pull requests.