docs(README.md): add detailed information about the purpose of the project and the resources it creates
This commit adds a detailed description of the project and the resources
it creates. It explains that the project creates a GCS bucket through
terraform cdktf API, a GC logging bucket, and a GC sink that routes the
logs to the logging bucket. It also provides information about the
prerequisites for running the project, including installing terraform,
learning CDK for terraform, and learning features of GCP. Additionally,
it provides information about the required IAM roles and permissions for
the service account used to run terraform and how to download the json
formatted service account key. Finally, it provides information about
how to run the commands to deploy the application.